requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1
python-multipart==0.0.6
pydub==0.25.1
soundfile==0.12.1
pyannote.audio==3.1.1
torch==2.1.0
torchaudio==2.1.0
numpy==1.24.3
scipy==1.11.3
requests==2.31.0
python-dotenv==1.0.0